Adding citation agents
