+++ f_cache_temp_july/cache.json
@@ -1,1 +1,1 @@
-{"https://www.weather.gov/bgm/climatePlots2024SYR": {"obtained_on": "2024-02-13T23:47:02.419", "last_accessed": "2024-02-13T23:47:02.420", "accessed": 1, "uri_md5": "1835e7fd506978336b5debdf5a4ecfac"}, "https://www.ncei.noaa.gov/access/monitoring/monthly-report/national/": {"obtained_on": "2024-02-13T23:47:07.122", "last_accessed": "2024-02-13T23:47:07.122", "accessed": 1, "uri_md5": "bda8ec3e9faec66043baa4392023613a"}, "https://www.weather.gov/cae/cliplotcae2024": {"obtained_on": "2024-02-13T23:47:07.509", "last_accessed": "2024-02-13T23:47:07.510", "accessed": 1, "uri_md5": "7cc71ff22249e1538490734ff1731e9e"}, "https://nsidc.org/arcticseaicenews/": {"obtained_on": "2024-02-13T23:47:08.712", "last_accessed": "2024-02-13T23:47:08.713", "accessed": 1, "uri_md5": "e70c9159c6d3fa96d7f0e86dcd881518"}, "https://water.weather.gov/precip/": {"obtained_on": "2024-02-13T23:47:09.099", "last_accessed": "2024-02-13T23:47:09.100", "accessed": 1, "uri_md5": "a53c14966ee034a4188eaa481365a697"}, "https://data.giss.nasa.gov/gistemp/": {"obtained_on": "2024-02-13T23:47:09.724", "last_accessed": "2024-02-13T23:47:09.725", "accessed": 1, "uri_md5": "f909a9458110a71d92456a81afb8db7d"}, "https://climatesciences.jpl.nasa.gov/events/summer-school/about/": {"obtained_on": "2024-02-13T23:47:09.961", "last_accessed": "2024-02-13T23:47:09.962", "accessed": 1, "uri_md5": "b0b428218046fc00ac2b47392c13e4c4"}, "https://www.oregon.gov/odot/dmv/pages/offices/closures.aspx": {"obtained_on": "2024-02-13T23:47:10.639", "last_accessed": "2024-02-13T23:47:10.640", "accessed": 1, "uri_md5": "d012bb99abbd7686708b75d56c555c73"}}
+{"https://www.ncei.noaa.gov/access/monitoring/monthly-report/global/": {"obtained_on": "2024-02-14T19:03:17.279", "last_accessed": "2024-02-14T19:03:17.279", "accessed": 1, "uri_md5": "135edb862f59ec662a3ecb8e1726e3c5"}, "https://www.weather.gov/bgm/climatePlots2024SYR": {"obtained_on": "2024-02-14T19:03:17.432", "last_accessed": "2024-02-14T19:03:17.433", "accessed": 1, "uri_md5": "1835e7fd506978336b5debdf5a4ecfac"}, "https://www.ncei.noaa.gov/access/monitoring/monthly-report/national": {"obtained_on": "2024-02-14T19:03:21.747", "last_accessed": "2024-02-14T19:03:21.747", "accessed": 1, "uri_md5": "5276d6dfe312dd7842049eb3f4631032"}, "https://www.weather.gov/cae/cliplotcae2024": {"obtained_on": "2024-02-14T19:03:21.942", "last_accessed": "2024-02-14T19:03:21.943", "accessed": 1, "uri_md5": "7cc71ff22249e1538490734ff1731e9e"}, "https://nsidc.org/arcticseaicenews/": {"obtained_on": "2024-02-14T19:03:24.000", "last_accessed": "2024-02-14T19:03:24.001", "accessed": 1, "uri_md5": "e70c9159c6d3fa96d7f0e86dcd881518"}, "https://water.weather.gov/precip/": {"obtained_on": "2024-02-14T19:03:24.684", "last_accessed": "2024-02-14T19:03:24.686", "accessed": 1, "uri_md5": "a53c14966ee034a4188eaa481365a697"}, "https://data.giss.nasa.gov/gistemp/": {"obtained_on": "2024-02-14T19:03:55.740", "last_accessed": "2024-02-14T19:03:55.742", "accessed": 1, "uri_md5": "f909a9458110a71d92456a81afb8db7d"}, "https://climatesciences.jpl.nasa.gov/events/summer-school/about/": {"obtained_on": "2024-02-14T19:03:55.929", "last_accessed": "2024-02-14T19:03:55.931", "accessed": 1, "uri_md5": "b0b428218046fc00ac2b47392c13e4c4"}, "http://www.climate.gov/news-features/understanding-climate/us-climate-outlook-january-2024": {"obtained_on": "2024-02-14T19:06:25.726", "last_accessed": "2024-02-14T19:06:25.727", "accessed": 1, "uri_md5": "3513bd2b47a49d93289e4e237c94fc41"}, "https://www.noaa.gov/news-release/us-winter-outlook-wetter-south-warmer-north": {"obtained_on": "2024-02-14T19:06:26.009", "last_accessed": "2024-02-14T19:06:26.011", "accessed": 1, "uri_md5": "a848fadc1d642cb43434c928ceba53dd"}, "https://www.weather.gov/mpx/mspclimate": {"obtained_on": "2024-02-14T19:06:26.286", "last_accessed": "2024-02-14T19:06:26.288", "accessed": 1, "uri_md5": "4e44f09a88e2d28c3f57229ef2606b26"}, "https://www.ncei.noaa.gov/news/groundhog-day-forecasts-and-climate-history": {"obtained_on": "2024-02-14T19:06:27.303", "last_accessed": "2024-02-14T19:06:27.304", "accessed": 1, "uri_md5": "d4058e47af2299f7ebb02de056c2688a"}, "https://www.weather.gov/ilm/WinterOutlook2023-2024": {"obtained_on": "2024-02-14T19:06:27.662", "last_accessed": "2024-02-14T19:06:27.663", "accessed": 1, "uri_md5": "de3c760a4ac3f889b557c315f3f80177"}, "https://www.cpc.ncep.noaa.gov/": {"obtained_on": "2024-02-14T19:06:28.186", "last_accessed": "2024-02-14T19:06:28.187", "accessed": 1, "uri_md5": "a73e2ed84a766ad0ddc9b08960d76ded"}, "https://www.cpc.ncep.noaa.gov/products/predictions/long_range/": {"obtained_on": "2024-02-14T19:06:28.693", "last_accessed": "2024-02-14T19:06:28.695", "accessed": 1, "uri_md5": "8860ce134ff46d5f900caa1b74cf9dd1"}, "https://www.wpc.ncep.noaa.gov/": {"obtained_on": "2024-02-14T19:06:29.607", "last_accessed": "2024-02-14T19:06:29.608", "accessed": 1, "uri_md5": "cb6368f9aae366a603d9385d6a08108b"}, "https://spectrumlocalnews.com/tx/south-texas-el-paso/weather/2023/10/29/winter-2023-2024-outlook-for-texas": {"obtained_on": "2024-02-14T19:06:30.104", "last_accessed": "2024-02-14T19:06:30.106", "accessed": 1, "uri_md5": "e05ac30ffe31ac48ac89467879adee2b"}}
