+++ f_cache_temp_july/original/e70c9159c6d3fa96d7f0e86dcd881518
@@ -252,7 +252,7 @@ height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
 <!--
 						<a href="https://nsidc.org/arcticseaicenews/">
 				                    <div id="home-feature">
-					<img src="https://nsidc.org/arcticseaicenews/wp-content/themes/asina-2012/images/headers/hanoi.jpg" width="1100" height="456" alt="" />
+					<img src="https://nsidc.org/arcticseaicenews/wp-content/themes/asina-2012/images/headers/willow.jpg" width="1100" height="456" alt="" />
                     </div>
 							</a>
 			
@@ -278,7 +278,7 @@ height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
         		                	<div id="daily-image">
 						<article id="post-1" class="indexed post-1 post type-post status-publish format-image sticky hentry category-daily-image post_format-post-format-image">
 			            <header class="entry-header">
-				<h2 class="entry-title"><a href="https://nsidc.org/arcticseaicenews/2012/05/daily-image/" title="Permalink to Daily Image Update" rel="bookmark">Daily Image Update</a> <!-- DATE HERE NEEDS TO BE YESTERDAY BECAUSE THERE'S A DAY LAG TIME IN THE POSTING Tue., February 13, 2024--></h2>
+				<h2 class="entry-title"><a href="https://nsidc.org/arcticseaicenews/2012/05/daily-image/" title="Permalink to Daily Image Update" rel="bookmark">Daily Image Update</a> <!-- DATE HERE NEEDS TO BE YESTERDAY BECAUSE THERE'S A DAY LAG TIME IN THE POSTING Wed., February 14, 2024--></h2>
              </header>
 		<div class="entry-content">
 			<div id="asina-overview-text" style="color: white;margin-top: 11px">Read scientific analysis on Arctic sea ice conditions. We provide an update during the first week of each month, or more frequently as conditions warrant.</div>
