+++ f_cache_temp_july/parsed/4e44f09a88e2d28c3f57229ef2606b26
@@ -1,0 +1,297 @@
+HOME
+
+
+
+
+FORECAST 
+   Local   Graphical   Aviation   Marine   Rivers and Lakes   Hurricanes   Severe Weather   Fire Weather   Sun/Moon   Long Range Forecasts   Climate Prediction   Space Weather  
+
+
+PAST WEATHER 
+   Past Weather   Heating/Cooling Days   Monthly Temperatures   Records   Astronomical Data  
+
+
+SAFETY 
+   Tsunamis   Floods   Beach Hazards   Wildfire   Cold   Tornadoes   Air Quality   Fog   Heat   Hurricanes   Lightning   Safe Boating   Rip Currents   Thunderstorms   Space Weather   Sun (Ultraviolet Radiation)   Safety Campaigns   Wind   Drought   Winter Weather  
+
+
+INFORMATION 
+   Wireless Emergency Alerts   Brochures   Weather-Ready Nation   Cooperative Observers   Daily Briefing   Damage/Fatality/Injury Statistics   Forecast Models   GIS Data Portal   NOAA Weather Radio   Publications   SKYWARN Storm Spotters   StormReady   TsunamiReady   Service Change Notices  
+
+
+EDUCATION 
+   Be A Force of Nature   NWS Education Home  
+
+
+NEWS 
+   NWS News   Events   Pubs/Brochures/Booklets    NWS Media Contacts  
+
+
+SEARCH 
+
+
+
+
+
+
+Search For
+
+
+
+
+NWS
+
+All NOAA
+
+
+
+
+
+
+
+
+
+ABOUT 
+   About NWS   Organization   Strategic Plan   Commitment to Diversity   For NWS Employees   International   National Centers   Careers   Contact Us   Glossary   Social Media  
+
+
+
+
+
+
+
+
+Local forecast by "City, St" or ZIP code
+
+
+
+Sorry, the location you searched for was not found. Please try another search.
+Multiple locations were found. Please select one of the following:
+
+
+
+Location Help
+
+
+
+
+
+
+
+
+News Headlines
+
+
+
+
+Latest Snowfall Reports
+Record January Thaw Still Going in Mid-February
+The Winter of 2023-2024
+Now Accepting Applications for Summer Student Volunteers
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+Customize Your Weather.gov
+
+
+
+Enter Your City, ST or ZIP Code
+
+Remember Me
+
+
+
+Privacy Policy
+
+
+
+
+LOADING...
+
+
+
+
+  Twin Cities, MN Weather Forecast Office   Past Weather Information for the Twin Cities  Weather.gov > Twin Cities, MN > Past Weather Information for the Twin Cities         
+     Current Hazards     Storm and Precipitation Reports   Outlooks   Submit a Storm Report   Experimental Graphical Hazardous Weather Outlook       Current Conditions     Satellite   Snowfall Analysis   Observations   Local Storm Reports   Local Snowfall Reports   Local Ice Accumulation Reports   Observed Precipitation       Radar     Local Enhanced Radar   Local Standard Radar (low bandwidth)   Regional Standard Radar (low bandwidth)        Forecasts     Hourly Forecasts   Activity Planner   User Defined Area Forecast   Fire Weather   Winter Weather   Forecast Discussion       Rivers and Lakes     Climate and Past Weather     Local   Drought   NOAA Climate Service   Event Summaries       Local Programs     Blog   Publications   Winter Weather   Aviation Weather   Graphical Hazardous Weather Outlook   Wet Bulb Globe Temperature   Student Opportunities      
+
+
+  
+
+
+
+Daily Weather Observations for the Twin Cities
+(1995 to Current)
+  JanuaryFebruaryMarchAprilMayJuneJulyAugustSeptemberOctoberNovemberDecember   
+
+
+
+Monthly Temperature Graphs for the Twin Cities
+(2006 to Current)
+  JanuaryFebruaryMarchAprilMayJuneJulyAugustSeptemberOctoberNovemberDecember   
+
+
+
+Hourly Weather Conditions for the Past 3 Days
+Yesterday's Official Weather Records
+
+
+
+ 
+
+Other Links
+
+Minneapolis/St Paul Metro Area Climate Page (MN State Climatologist)
+Minnesota State Climatologist
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+  Follow us on Twitter
+                		
+
+
+
+
+
+  Follow us on Facebook
+                		
+
+
+
+
+
+  Follow us on YouTube
+            			
+
+
+
+
+
+
+  MPX RSS Feed
+						
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+Warnings/Hazards
+
+Experimental Graphical Hazardous Weather Outlook
+
+
+
+
+
+Weather  Forecasts
+
+Snowfall/Winter Weather
+Aviation Forecasts
+Fire Weather
+Wet Bulb Globe Temperature
+
+
+
+
+
+Current Weather Conditions
+
+Current Weather Observations
+Yesterday's Temperatures and Precipitation
+Current Radar
+
+
+
+
+
+Past Weather Records
+
+On This Day in Weather History...
+Twin Cities Weather Records
+St Cloud Weather Records
+Eau Claire Weather Records
+Year-By-Year Minnesota Tornadoes
+
+
+
+
+
+Weather Safety
+
+NOAA Weather Radio
+Skywarn
+Spotter Activation
+
+
+
+
+
+Office Activities
+
+Cooperative Weather Observer
+Student Opportunities
+
+
+
+
+
+
+
+
+
+
+
+
+
+US Dept of Commerce
+National Oceanic and Atmospheric Administration
+National Weather Service
+                        Twin Cities, MN1733 Lake Drive WestChanhassen, MN 55317-8581952-361-6670Comments? Questions? Please Contact Us. 
+
+
+
+Disclaimer
+Information Quality
+Help
+Glossary
+
+
+Privacy Policy
+Freedom of Information Act (FOIA)
+About Us
+Career Opportunities
