+++ f_cache_temp_july/parsed/5276d6dfe312dd7842049eb3f4631032
@@ -17,7 +17,7 @@ January 2024 National Climate Report
 
 
 HomeClimate MonitoringJanuary 2024 National Climate Report
-January Global Release: Wed, 14 Feb 2024, 11:00 AM EST
+February U.S. Release: Fri, 8 Mar 2024, 11:00 AM EST
 
 
 
@@ -213,7 +213,7 @@ Notable Weather
 Atmospheric River Causes Flash Flooding in San Diego, California: An atmospheric river hit southern California from January 21st to January 22nd dropping 2 to 3 inches in San Diego prompting the mayor to declare a state of emergency.  The rain caused severe flash flooding leading to many water rescues, closed roadways, and massive property damage.  Multiple landslides and rockslides were also caused by the excess water.
 For more information, please go to the Western Regional Climate Center Home Page.
 
-Citing This ReportNOAA National Centers for Environmental Information, Monthly National Climate Report for January 2024, published online February 2024, retrieved on February 13, 2024 from https://www.ncei.noaa.gov/access/monitoring/monthly-report/national/202401.Metadatahttps://www.ncei.noaa.gov/access/metadata/landing-page/bin/iso?id=gov.noaa.ncdc:C00674
+Citing This ReportNOAA National Centers for Environmental Information, Monthly National Climate Report for January 2024, published online February 2024, retrieved on February 14, 2024 from https://www.ncei.noaa.gov/access/monitoring/monthly-report/national/202401.Metadatahttps://www.ncei.noaa.gov/access/metadata/landing-page/bin/iso?id=gov.noaa.ncdc:C00674
