+++ f_cache_temp_july/parsed/7cc71ff22249e1538490734ff1731e9e
@@ -401,9 +401,9 @@ Mesonet Data
 
 Radar
 
+Charlotte TDWR 
 Nationwide 
 Columbia Radar 
-Charlotte TDWR
