+++ f_cache_temp_july/parsed/8860ce134ff46d5f900caa1b74cf9dd1
@@ -1,0 +1,445 @@
+www.nws.noaa.gov 
+				
+
+
+
+
+
+
+
+
+
+ 
+
+
+
+ 
+
+
+
+
+
+
+
+Site Map
+
+
+News
+
+
+Organization
+
+
+
+Search 
+
+
+
+
+
+
+
+
+ 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+ 
+CPC Search 
+ 
+
+
+
+
+
+
+Map Explanations
+								   Official Fcsts
+								   Fcst Tools
+Past Forecasts
+								   Archives/Verifications
+								   Seasonal Postmortem
+About Us
+								   Our Mission
+								   Who We Are
+Contact Us
+								   CPC Information
+								   CPC Web Team
+
+
+ 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+HOME > 
+								Outlook Maps > 
+								Monthly to Seasonal Outlooks > 
+								Seasonal Outlooks
+								
+
+
+
+
+
+
+
+
+
+
+
+
+
+OFFICIAL Forecasts
+
+
+
+
+
+Click below for archives of past outlooks (data & graphics), verifications of past outlooks, and GIS data for current outlooks
+
+Archives
+Verifications
+GIS Data
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+Text-FormatDiscussions
+Prognostic Discussion
+Tools Discussion
+30-& 90-Day Hawaiian Discussion
+
+Briefing Materials (updated as new information arrives)
+
+
+Official 90-day Outlooks are issued once each month near mid-month at 8:30am Eastern Time. Please consult the schedule of 30 & 90-day outlooksfor exact release dates.
+
+
+
+
+
+
+
+
+
+
+
+OUTLOOKS
+
+(click title for maps)
+
+
+Norm
+
+
+Prob of Exceed
+(Interactive)
+
+
+
+T
+P
+
+
+
+0.5 Month Outlook for Feb-Mar-Apr 2024
+
+
+N
+
+
+T
+
+
+P
+
+
+
+1.5 Month Outlook for Mar-Apr-May 2024
+
+
+N
+
+
+T
+
+
+P
+
+
+
+2.5 Month Outlook for Apr-May-Jun 2024
+
+
+N
+
+
+T
+
+
+P
+
+
+
+3.5 Month Outlook for May-Jun-Jul 2024
+
+
+N
+
+
+T
+
+
+P
+
+
+
+4.5 Month Outlook for Jun-Jul-Aug 2024
+
+
+N
+
+
+T
+
+
+P
+
+
+
+5.5 Month Outlook for Jul-Aug-Sep 2024
+
+
+N
+
+
+T
+
+
+P
+
+
+
+6.5 Month Outlook for Aug-Sep-Oct 2024
+
+
+N
+
+
+T
+
+
+P
+
+
+
+7.5 Month Outlook for Sep-Oct-Nov 2024
+
+
+N
+
+
+T
+
+
+P
+
+
+
+8.5 Month Outlook for Oct-Nov-Dec 2024
+
+
+N
+
+
+T
+
+
+P
+
+
+
+9.5 Month Outlook for Nov-Dec-Jan 2024 - 25
+
+
+N
+
+
+T
+
+
+P
+
+
+
+10.5 Month Outlook for Dec-Jan-Feb 2024 - 25
+
+
+N
+
+
+T
+
+
+P
+
+
+
+11.5 Month Outlook for Jan-Feb-Mar 2025
+
+
+N
+
+
+T
+
+
+P
+
+
+
+12.5 Month Outlook for Feb-Mar-Apr 2025
+
+
+N
+
+
+T
+
+
+P
+
+
+
+
+
+
+Color Maps
+
+
+(click title for maps)
+
+
+
+
+
+
+Individual Outlooks
+
+
+
+
+
+Outlooks for February 2024 & Feb-Mar-Apr 2024
+
+
+
+
+
+All Temperature Outlooks
+
+
+
+
+
+
+All Precipitation Outlooks
+
+
+
+
+
+
+
+
+
+
+
+
+
+Email to Forecasters
+
+
+Email to Webmasters
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+NOAA/
+National Weather Service
+National Centers for Environmental Prediction
+														Climate Prediction Center
+														5830 University Research Court
+														College Park, Maryland 20740
+														Page Author: Climate Prediction Center Internet Team
+														Page last modified: December 14, 2005
+													
+
+Disclaimer
+
+Information Quality
+
+Credits
+
+Glossary
+
+
+
+Privacy Policy
+Freedom of Information Act (FOIA)
+
+About Us
+
+Career Opportunities
