+++ f_cache_temp_july/parsed/a53c14966ee034a4188eaa481365a697
@@ -274,6 +274,19 @@ Daily
 
 Time Range:
 
+February 14, 2024 - Today
+February 14, 2024 - Last 7 Days
+February 14, 2024 - Last 14 Days
+February 14, 2024 - Last 30 Days
+February 14, 2024 - Last 60 Days
+February 14, 2024 - Last 90 Days
+February 14, 2024 - Last 180 Days
+February 14, 2024 - Last 365 Days
+February 14, 2024 - Month to Date
+February 14, 2024 - Year to Date
+February 14, 2024 - Water Year to Date (Oct. 1)
+
+
 February 13, 2024 - Today
 February 13, 2024 - Last 7 Days
 February 13, 2024 - Last 14 Days
@@ -285,19 +298,6 @@ February 13, 2024 - Last 365 Days
 February 13, 2024 - Month to Date
 February 13, 2024 - Year to Date
 February 13, 2024 - Water Year to Date (Oct. 1)
-
-
-February 12, 2024 - Today
-February 12, 2024 - Last 7 Days
-February 12, 2024 - Last 14 Days
-February 12, 2024 - Last 30 Days
-February 12, 2024 - Last 60 Days
-February 12, 2024 - Last 90 Days
-February 12, 2024 - Last 180 Days
-February 12, 2024 - Last 365 Days
-February 12, 2024 - Month to Date
-February 12, 2024 - Year to Date
-February 12, 2024 - Water Year to Date (Oct. 1)
