+++ f_cache_temp_july/parsed/a73e2ed84a766ad0ddc9b08960d76ded
@@ -1,0 +1,603 @@
+www.nws.noaa.gov 
+			
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+				 
+			
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+Home
+
+
+Site Map
+
+
+News
+
+
+Organization
+
+
+
+  Enter Search Term(s): 
+
+
+
+
+
+
+
+
+
+
+
+ 
+
+
+
+
+
+
+
+
+
+
+
+Search the CPC
+
+
+
+
+
+
+
+
+ 
+
+   NCEP Quarterly       Newsletter
+
+ 
+
+ Climate-Weather
+		  El Niño/La Niña
+		  MJO
+		  Blocking
+		  AAO, AO, NAO, PNA
+		  Climatology
+		  Global Monsoons
+
+ 
+
+Expert Assessments
+		   Index
+		   Products
+
+ 
+
+Outlooks (Forecasts)
+		   Index
+		   Products
+		   Verification
+
+ 
+
+Monitoring and Data
+		   Index
+		   Products
+
+ 
+
+Crosscutting Themes
+		   Ocean Climate
+		   Stratosphere
+		   Pacific Islands
+		   International
+		        Desks
+
+ 
+
+Partnerships
+		   Climate.gov
+		   Climate Test Bed
+		   NOAA
+		   Interagency
+		   JAWF
+		   USAID
+		   Alaska
+		   Others
+
+
+
+
+Outreach
+		   Educational
+		   Climate Glossary
+		   Meetings
+		   CPC Seminars
+		   Publications
+
+ 
+
+About Us
+		   Vision & Mission
+		   Who We Are
+
+ 
+
+Contact Us
+		   CPC Information
+		   CPC Web Team
+
+ 
+ 
+
+
+
+
+
+ 
+
+
+
+
+ 
+
+
+
+
+
+
+
+
+DOC
+
+
+NOAA
+
+
+NWS
+
+
+NCEP Centers:
+
+
+AWC
+
+
+CPC
+
+
+EMC
+
+
+NCO
+
+
+NHC
+
+
+OPC
+
+
+SPC
+
+
+SWPC
+
+
+WPC
+
+
+
+
+
+
+
+ 
+
+
+Climate News
+
+
+
+
+
+A transition from El Niño to ENSO-neutral is likely by April-June 2024 (79% chance), with increasing odds of La Niña developing in June-August 2024 (55% chance) (8 Feb 2024)
+
+
+
+
+48th Climate Diagnostics and Prediction Workshop Announcement (6 Apr 2023)
+
+
+
+
+
+
+
+
+
+
+
+
+Click on product title to go to product page. Move cursor over product parameter name to display the graphic -- click to enlarge. Links to these same products are also available below.
+
+
+
+
+
+
+
+
+
+
+
+
+6-10 Day Outlook  
+																		 (Interactive)
+Temperature
+																		                  
+																		 Precipitation
+
+ 
+
+One Month Outlook  
+																		(Interactive)
+Temperature
+																		                   
+																		Precipitation
+
+
+
+ 
+
+
+
+8-14 Day Outlook  
+																		 (Interactive)
+Temperature
+																		                  
+																		 Precipitation
+
+ 
+
+Three Month Outlook  
+																		(Interactive)
+Temperature
+																		                  
+																		Precipitation
+
+
+
+ 
+
+
+
+Week 3-4 Outlooks  
+Temperature
+																		                  
+																		 Exp. Precipitation
+
+ 
+
+8-14 Day U.S. Hazards Outlook
+
+Composite      
+															Probabilistic:   Temp   
+																		Precip   
+
+Snow   
+																		Wind
+
+
+
+ 
+
+
+
+
+U.S. Drought Information
+Monitor   
+																		Monthly Outlook   
+																		Seasonal Outlook
+
+ 
+
+Global Tropics Hazards Outlook  
+Weeks 2 and 3
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+Local Climate, Water & Weather Topics
+
+
+
+
+U.S. Hazards Outlook
+U.S. Drought Information
+ENSO Diagnostic Discussion
+Climate-Weather
+Expert Assessments Products
+Expert Assessments Index
+
+
+Outlooks Products
+Outlooks Index
+Monitoring and Data Products
+Monitoring and Data Index
+El Niño/La Niña
+Stratosphere
+
+
+Pacific Islands
+International
+Partnerships
+Outreach
+CPC Information
+Contact Us
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+NOAA/
+National Weather Service
+NOAA Center for Weather and Climate Prediction 
+						Climate Prediction Center
+						5830 University Research Court 
+						College Park, Maryland 20740 
+						Page Author: Climate Prediction Center Internet Team
+						Page last modified: Oct 13 2023			
+
+
+Disclaimer
+
+Information Quality
+
+Credits
+
+Glossary
+
+
+
+Privacy Policy
+
+Freedom of Information Act (FOIA)
+
+About Us
+
+Career Opportunities
