--- f_cache_temp_july/parsed/d012bb99abbd7686708b75d56c555c73
@@ -1,356 +1,0 @@
-Skip to main content
-
-
-
-
-
-
-
-
-
-An official website of the State of Oregon
-                         Learn
-How you know »
-(how to identify a Oregon.gov website)
-
-An official website of the State of Oregon »
-                    
-
-
-
-
-
-
-
-Translate this site into other Languages
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-Toggle Main Menu
-
-
-
-
-
-
-
-		Main Navigation
-	
-HomeLicenses & ID CardsVehiclesFormsRecordsOfficesMost Popular
-
-
-
-
-
-
-
-
-
-
-close
-
-
-Search
-
-
-
-Submit
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-                                        <div class='or-header-logo'><span class='sr-only'></span></div>
-
-                                    
-
-
-
-
-
-                                        Oregon Driver & Motor Vehicle Services
-                                    
-
-
-
-
-
-
-You are here:
-
-Oregon Department of TransportationOregon Driver & Motor Vehicle ServicesDMV Offices - Statewide Holiday Closures
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-                                
-                                DMV Offices - Statewide Holiday Closures
-                            
-
- Site Navigation
-
-
-
-
-
-
-
-
-
-About UsLicenses & ID CardsVehiclesFormsRecords50plusTeenFAQsDMV2UDMV Job Opportunities
-
-
-
-
-
-
-
-
-
- 
-   During inclement weather, or state emergencies, please see the Department of Administrative Services Building Closure Notices for State OfficesAll DMV offices, including DMV Headquarters and the Business Regulation office in Salem, will be closed on the dates listed below due to holidays. 
-
-Note: Please see the 
-      Web page for your target DMV office for unscheduled office closures.2024 ClosuresState agencies will be closed on the following days in 2024:New Years Day observed - Monday, Jan. 1, 2024Martin Luther King Jr. Day - Monday, Jan. 15, 2024President's Day - Monday, Feb. 19, 2024Memorial Day - Monday, May 27, 2024Juneteenth - Wednesday, June 19, 2024Independence Day - Thursday, July 4, 2024Labor Day - Monday, Sept. 2, 2024Veterans Day observed - Monday, Nov. 11, 2024Thanksgiving Day - Thursday, Nov. 28, 2024Day after Thanksgiving - Friday, Nov. 29, 2024Christmas Day - Wednesday, Dec. 25, 2024
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-	Footer
-
-
-
-			Links
-		
-
-TripCheck (24/7 travel info)
-Message from the Division Administrator
-FAQs Nondiscrimination and Equal Access
-Accessibility at ODOT
-Site Index
-Website Feedback
-
-
-
-
-
-
-
-			Social Media
-		
-
-
-
-Like ODOT on Facebook
-
-
-
-
-Follow ODOT on Twitter
-
-
-
-
-See ODOT on Flickr
-
-
-
-
-Watch ODOT on Youtube
-
-
-
-
-
-			Contact Us
-		
-
-Oregon DMV
-
-1905 Lana Ave NE
-Salem, OR 97314
-
-
- 1-503-945-5000
- Media Contacts
- Contact DMV
-
-
-
-
-			About Oregon
-		
-
-
-
-
- Back to Top
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-×
-
-How to recognize an official Oregon website
-
-
-
-
-.gov
-
-Official websites use .gov
-A .gov website belongs to an official government organization in the United States.
-
-
-
-
-
-Secure .gov websites use HTTPS
-A lock icon ( ) or https:// means you’ve safely connected to the .gov website.
-
-
-
-Only share sensitive information on official, secure websites.
-Close
-
-
-
-
-
-
-
-Hidden Submit
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-Your browser is out-of-date! It has known security flaws and may not display all features of this and other websites. Learn how
-×
