+++ f_cache_temp_july/parsed/d4058e47af2299f7ebb02de056c2688a
@@ -1,0 +1,481 @@
+Skip to main content
+
+
+
+
+
+
+
+
+
+
+An official website of the United States government
+Here's how you know
+
+
+Here's how you know
+
+
+
+
+
+
+
+
+
+The .gov means it’s official.
+
+              Federal government websites often end in .gov or .mil. Before sharing sensitive information, make sure you’re on a federal government site.
+            
+
+
+
+
+
+
+The site is secure.
+
+              The https:// ensures that you are connecting to the official website and that any information you provide is encrypted and transmitted securely.
+            
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+MENU
+
+
+
+
+
+
+
+
+
+Home
+
+
+Products
+
+
+Services
+
+
+Access
+
+
+Archive
+
+
+Certification
+
+
+Maps
+
+
+Monitoring
+
+
+Online Store
+
+
+Regional
+
+
+State of the Services
+
+
+World Data System
+
+
+
+
+Resources
+
+
+Alerts
+
+
+Change Notices
+
+
+Education
+
+
+Metadata
+
+
+Quick Links
+
+
+
+
+News
+
+
+BAMS
+
+
+Publications
+
+
+
+
+Contact
+
+
+About
+
+
+Locations
+
+
+Our Impact
+
+
+
+
+
+
+
+
+
+
+
+
+Enter Search Term(s):
+
+
+
+                Search
+              
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+ Please note: Due to scheduled maintenance, many NCEI systems will be unavailable February 13th, 8:00 AM ET - February 14th, 10:00 PM ET. We apologize for any inconvenience.
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+Home
+
+
+News
+
+
+                  Groundhog Day Forecasts and Climate History
+              
+
+
+
+
+Groundhog Day Forecasts and Climate History
+
+
+
+
+
+
+
+
+
+
+
+ 
+
+
+ 
+    Courtesy of Gettyimages
+    
+
+
+
+Every February 2, a crowd of thousands gathers at Gobbler’s Knob in Punxsutawney, Pennsylvania, to await a special forecast from a groundhog named Phil. According to legend, if the 20-pound groundhog emerges and sees his shadow, the United States can expect six more weeks of winter weather. But, if Phil doesn’t see his shadow, we can expect warmer temperatures and the arrival of an early spring.Even though he’s been forecasting since 1887, Phil’s track record for the entire country isn’t perfect. To determine just how accurate he is, we’ve compared U.S. national temperatures with Phil’s forecasts. On average, Phil has gotten it right 30% of the time over the past 10 years.Phil’s 2023 ForecastIn 2023, Phil forecast a "long winter" when he saw his shadow and predicted an additional six weeks of wintry temperatures. In fact, the contiguous United States saw above average temperatures in February and slightly below average temperatures in March of last year. Phil was 50/50 on his forecast.The average contiguous U.S. temperature during February 2023 was 36.5°F. It was 2.6°F above the 20th century average. 2023 was the 102nd-coldest February in the 128-year period of record.February 2023 was above average across much of the eastern half of the contiguous U.S., including parts of the northern High Plains, with below-average temperatures from the central Rockies to the West Coast. The Alaska statewide February temperature was 6.0°F, which is 1.2°F above the long-term average. March 2023 was cooler than average in the contiguous U.S. The average temperature was 40.6°F, 0.9°F below the 20th-century average. Generally, temperatures were above average from the southern Plains to New England and in parts of the Great Lakes, with below-average temperatures from the northern Plains to the West Coast. Alaska had an average temperature of 13.5°F, which is 2.7°F above the long-term average.Phil’s First ForecastIn 1887, when he made his debut as the official groundhog forecaster for the entire country, Phil saw his shadow. His first prediction of six more weeks of winter was accurate for a few regions, but it came up short for several others.According to the February 1887 Monthly Weather Review, the Northeast, Great Lakes region, and West saw temperatures well below normal. The Southeast and Gulf states saw temperatures well above normal during the month. And, according to the March 1887 Monthly Weather Review, the Northeast, Great Lakes region, Ohio Valley, and Southeast saw temperatures well below normal. Areas west of the Mississippi River valley saw temperatures above normal.Predicting the Arrival of Spring Is DifficultPredicting the arrival of springtime for an entire country—especially one with varied regional climates like the United States—isn’t easy! Phil’s track record is evidence of that. However, if you’re interested in doing your own analysis for your region, check out our Climate at a Glance tool to access historical U.S. monthly temperature data. More of Phil’s past predictions are also available from the Punxsutawney Groundhog Club.For an overview of some fun facts about Groundhog Day and the accuracy of the furry forecaster, check out our infographic. To see the latest climate outlooks, visit NOAA’s Climate Prediction Center. For the current weather forecast in your area, check out your local National Weather Service forecast office. 
+
+
+
+Published
+    February 2, 2017
+
+
+Updated Date
+ 
+    February 2, 2024
+    
+
+
+Now includes the latest analysis and climate information for 2023.
+
+
+Related Links
+
+U.S. Climatological Rankings
+Climate at a Glance
+U.S. Monthly, Seasonal, and Annual Climate Reports
+U.S. Climate Normals
+
+
+
+Related Files
+
+February 1887 Monthly Weather Review Form
+
+March 1887 Monthly Weather Review Form
+
+
+
+
+Article Tags
+
+ 
+Education/Outreach
+
+
+ 
+Air Temperature
+
+
+ 
+Holidays
+
+
+ 
+Animals
+
+
+ 
+History/Folklore
+
+
+
+
+ 
+North America
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+Related News
+
+
+
+
+
+
+
+
+
+
+
+      February 9, 2024
+
+
+  
+
+U.S. Drought: Weekly Report for February 6, 2024
+
+
+
+Read More
+
+
+
+
+
+
+
+
+
+
+
+
+      February 2, 2024
+
+
+  
+
+Groundhog Day Forecasts and Climate History
+
+
+
+Read More
+
+
+
+
+
+
+
+
+
+
+
+
+      February 1, 2024
+
+
+  
+
+U.S. Drought: Weekly Report for January 30, 2024
+
+
+
+Read More
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+Return to top
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+Follow Us
+
+
+@NOAANCEI
+
+
+@NOAAData
+
+
+@NOAANCEI
+
+
+News Feed
+
+
+
+
+
+
+Contact Us
+
+
+Email by topic
+
+
+Phone: (828) 271-4800
+
+
+
+
+  
+
+
+
+
+
+
+Sitemap
+
+
+Privacy Policy
+
+
+Freedom of Information Act
+
+
+Information Quality
+
+
+Accessibility
+
+
+Disclaimer
+
+
+Take Our Survey
+
+
+Department of Commerce
+
+
+NOAA
+
+
+NESDIS
