+++ testagent.py
@@ -1,4 +1,8 @@
-from emergingtrajectories.agents2 import CitationScrapeAndPredictAgent, clean_citations
+from emergingtrajectories.citationagents import (
+    CitationScrapeAndPredictAgent,
+    CiteExtendScrapePredictAgent,
+    clean_citations,
+)
 from emergingtrajectories.knowledge import KnowledgeBaseFileCache
 
 import os
@@ -12,7 +16,7 @@ google_search_id = os.getenv("GOOGLE_SEARCH_ID")
 
 kb = KnowledgeBaseFileCache("f_cache_temp_july")
 
-agent_results = CitationScrapeAndPredictAgent(
+"""agent_results = CitationScrapeAndPredictAgent(
     openai_api_key,
     google_api_key,
     google_search_id,
@@ -20,10 +24,21 @@ agent_results = CitationScrapeAndPredictAgent(
     kb,
     8,
     et_api_key,
-    prediction_agent="[test] Citation + Web Scraper - July 2024 Temperature Anomalies"
+    prediction_agent="[test] Citation + Web Scraper - July 2024 Temperature Anomalies",
 )
 
-print(agent_results)
+print(agent_results)"""
 
-#text = clean_citations("Hi there 33 [ ], [44], how are you [21]?", {44:"abc", 21:"def"})
-#print(text)
+agent_results = CiteExtendScrapePredictAgent(
+    openai_api_key,
+    google_api_key,
+    google_search_id,
+    "Forecasts for 2024 temperature and weather records",
+    kb,
+    79,
+    et_api_key,
+    prediction_agent="[test, extended] Citation + Web Scraper - July 2024 Temperature Anomalies",
+)
+
+# text = clean_citations("Hi there 33 [ ], [44], how are you [21]?", {44:"abc", 21:"def"})
+# print(text)
